+++ src/commands/enroll.mjs
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+import { runEnrollCommand } from '../graph/enroll.mjs'
+runEnrollCommand()